+++ src/Tests/Unit/Public/Get-GMapPlaceDetail.Tests.ps1
@@ -76,11 +76,13 @@ InModuleScope 'pwshPlaces' {
                     $Uri | Should -BeLike '*formatted_phone_number*'
                     $Uri | Should -BeLike '*price_level*'
                     $Uri | Should -BeLike '*region=*'
+                    $Uri | Should -BeLike '*reviews_sort=newest*'
                 } -Verifiable
                 $getGMapPlaceDetailSplat = @{
                     PlaceID      = 'ChIJf9Yxhme9XIYRkXo-Bl62Q10'
                     Contact      = $true
                     Atmosphere   = $true
+                    ReviewSort   = 'Newest'
                     Language     = 'en'
                     RegionBias   = 'us'
                     GoogleAPIKey = $googleAPIKey